requirements.txt
https://github.com/Mike-Scarlet/pyrogram/archive/refs/heads/feature/custom_modify.zip
python-telegram-bot